+++ packages/core/src/components/AlertBanner/__stories__/AlertBanner.stories.tsx
@@ -5,6 +5,7 @@ import AlertBannerLink from "../AlertBannerLink/AlertBannerLink";
 import AlertBannerButton from "../AlertBannerButton/AlertBannerButton";
 import { createStoryMetaSettingsDecorator } from "../../../storybook/functions/createStoryMetaSettingsDecorator";
 import "./AlertBanner.stories.scss";
+import { Meta, StoryObj } from "@storybook/react";
 
 const metaSettings = createStoryMetaSettingsDecorator({
   component: AlertBanner
@@ -24,7 +25,7 @@ export default {
   component: AlertBanner,
   argTypes: metaSettings.argTypes,
   decorators: metaSettings.decorators
-};
+} satisfies Meta<typeof AlertBanner>;
 
 export const Overview = {
   render: alertBannerTemplate.bind({}),
@@ -35,7 +36,9 @@ export const Overview = {
   }
 };
 
-export const Types = {
+type Story = StoryObj<typeof AlertBanner>;
+
+export const Types: Story = {
   render: () => (
     <div className="monday-storybook-alert-banner_column-wrapper monday-storybook-alert-banner_big-container">
       <AlertBanner>
@@ -64,7 +67,7 @@ export const Types = {
   name: "Types"
 };
 
-export const AlertBannerWithButton = {
+export const AlertBannerWithButton: Story = {
   render: () => (
     <AlertBanner className="monday-storybook-alert-banner_big-container">
       <AlertBannerText text="Lorem ipsum dolor sit amet" />
@@ -75,7 +78,7 @@ export const AlertBannerWithButton = {
   name: "Alert Banner with button"
 };
 
-export const AlertBannerWithLink = {
+export const AlertBannerWithLink: Story = {
   render: () => (
     <AlertBanner className="monday-storybook-alert-banner_big-container">
       <AlertBannerText text="Alert banner message" />
@@ -86,7 +89,7 @@ export const AlertBannerWithLink = {
   name: "Alert Banner with link"
 };
 
-export const AlertBannerAsAnAnnouncement = {
+export const AlertBannerAsAnAnnouncement: Story = {
   render: () => (
     <AlertBanner backgroundColor="dark" className="monday-storybook-alert-banner_big-container">
       <AlertBannerText text="Join us at Elevate 2022" />
@@ -97,7 +100,7 @@ export const AlertBannerAsAnAnnouncement = {
   name: "Alert banner as an announcement"
 };
 
-export const AlertBannerAsAnOpportunityToUpgrade = {
+export const AlertBannerAsAnOpportunityToUpgrade: Story = {
   render: () => (
     <AlertBanner className="monday-storybook-alert-banner_big-container">
       <AlertBannerText text="7 days left on your monday CRM trial" />
@@ -108,7 +111,7 @@ export const AlertBannerAsAnOpportunityToUpgrade = {
   name: "Alert banner as an opportunity to upgrade"
 };
 
-export const OverflowText = {
+export const OverflowText: Story = {
   render: () => (
     <AlertBanner className="monday-storybook-alert-banner_small-container">
       <AlertBannerText text="This is a really long alert..." />