+++ packages/core/src/components/Dropdown/Dropdown.tsx
@@ -108,7 +108,8 @@ const Dropdown: VibeComponent<DropdownComponentProps, HTMLElement> & {
       popupsContainerSelector,
       filterOption,
       menuPosition = "absolute",
-      "data-testid": dataTestId
+      "data-testid": dataTestId,
+      withGroupDivider = false
     }: DropdownComponentProps,
     ref: React.ForwardedRef<HTMLElement>
   ) => {
@@ -159,7 +160,8 @@ const Dropdown: VibeComponent<DropdownComponentProps, HTMLElement> & {
         rtl,
         insideOverflowContainer,
         controlRef,
-        insideOverflowWithTransformContainer
+        insideOverflowWithTransformContainer,
+        withGroupDivider
       });
 
       type BaseStyles = typeof baseStyles;