+++ packages/core/src/components/Dropdown/Dropdown.types.ts
@@ -267,6 +267,7 @@ export interface DropdownComponentProps extends CustomMenuBaseProps, CustomOptio
   onClear?: () => void;
   popupsContainerSelector?: string;
   selectProps?: Record<string, string>;
+  withGroupDivider?: boolean;
 }
 
 export type DropdownProps = DropdownComponentProps;